Move serializeProxyProperties to Utilities
